fix(reader): train in fp32 -- fp16 parameters were the ln(5) plateau
Root cause of four failed training runs across three days. transformers 5.x
makes from_pretrained follow the CHECKPOINT's stored dtype; deberta-v3-base and
-large both ship fp16, so every training run in this project -- from the first --
loaded fp16 PARAMETERS and had AdamW update half-precision weights in place.
pyproject.toml pinned `transformers>=4.46` with no upper bound; under 4.x the
same call returned fp32.

An AdamW step is ~lr in magnitude, so the question is whether lr is
representable as a change to the weight. Measured over 20 steps at lr=2e-5 with
a constant-sign gradient: at w=0.03 fp16 moves 3.05e-04 vs fp32's 4.00e-04, and
at w >= 0.1 fp16 moves EXACTLY ZERO -- ULP grows with magnitude, and DeBERTa's
LayerNorm weights sit near 1.0. On the real recipe one step moved
classifier.weight by 1.072e-01 in fp16 against a correct 2.001e-05 in fp32.

Controlled test, 16 rows / 60 optimizer steps / one variable changed:
  fp16: loss 1.53 -> 4.32 -> 1.5687   parked at ln(5)=1.6094
  fp32: loss 1.68 -> 0.47 -> 0.1013   learns

Retrodictive, which is why it is trusted over the four earlier stories: it
explains ln(5) on both a T4 and a Blackwell card (a library default, not
hardware), why 3.3x more steps changed nothing, why source-matching changed
nothing, why inference was always fine (fp16 forward is fine -- the 0.761131
leaderboard score stands), and why the 0.5641 closed-book peak decayed and row 6
capped at 0.6086: small head weights are the one regime where fp16 still moves,
so the head learned while the encoder stayed frozen.

Ruled out first, and logged: the training pool's labels and context are sound
(known-good reader scores 0.8037 [0.7743,0.8320] on the TRAIN file vs 0.7970 on
eval), and our training input format costs nothing (0.7947). Truncation was a
red herring -- 416 vs 348 context tokens, a 16% haircut not 72%. Gradients were
always healthy and the optimizer always covered the right 77.2M params.

- dtype=torch.float32 on all 17 training-path loads; inference stays fp16
- assert_trainable_dtype/load_mc_model_for_training in reader/mc.py, inline
  equivalents in the Kaggle kernels (cannot import src/)
- tests/test_trainable_dtype.py: 10 tests on the mechanism itself, 3s
- hypothesis_gate.py check 0, running first: a passing DATA gate is not evidence
  training is possible, which is exactly how this survived the previous gate
- transformers pinned <6

Also recorded: on WSL2 exceeding VRAM raises "CUDA driver error: device not
ready", NOT torch.OutOfMemoryError, so the submission scripts' OOM->CPU fallback
would not fire.

+++ src/llmsci/gpu_guard.py
@@ -95,7 +95,7 @@ def probe_training_speed(
     committing to a run that might silently take 100x longer than expected.
     Frees its GPU memory before returning.
     """
-    probe_model = AutoModelForMultipleChoice.from_pretrained(model_name).to(device)
+    probe_model = AutoModelForMultipleChoice.from_pretrained(model_name, dtype=torch.float32).to(device)
     probe_model.train()
     probe_optimizer = torch.optim.AdamW(probe_model.parameters(), lr=lr, eps=eps)
